Add a basic example of outputting the nav
As a newcomer to Statamic it took me some time to understand how to output a basic nav once I'd created a couple of pages—this was mostly due to `include_home="true"` being an opt-in pattern that's not immediately obvious until I read the parameters carefully. I think having a couple of basic examples (with this parameter included) will really help newcomers like me.