Add honeypot field with fallback
If the name for the honeypot field is not explicitly set, Statamic checks for a field named `honeypot`. However, if not set, the `honeypot` variable is `null` instead of the default value `honeypot`. I would imagine that this implicit behavior would be somewhat better understood by the example with a fallback for the variable.