Update the description about valid variable names
The copy says that hyphens are allowed within a variable name but the regex does not reflected that and after a chat with @JohnathonKoster it turned out that hyphens at the end of a variable name is not allowed. I rephrased the paragraph and added the complete regex which describes a variable name.