Add lang option to config example on multi-site.md
A lang option was added to enable keeping a general locale but using a difference set of language files, this isn't currently ver obvious from the docs, but as a simple first step adding it into the example config could help. 

We could potentially show this as commented out with a note about it being optional?