Undocument DigitalOcean App Platform for the time being
Just ran through this guide, and it turns out it's pretty out-of-date now, especially with the fact it has a read-only filesystem which means you need to use the Eloquent Driver.

You also need to configure trusted proxies and know to move drivers to the database.

It's also pretty expensive compared to a traditional droplet, which would work fine for most sites.

With all that said, we're going to undocument this for now and write a generalised serverless guide in the future, see #1448.