Update image-manipulation.md
fix a papercut when configuring glide to use a CDN cache - despite the warning on line 214, I managed to overlook specifying   a distinct url and couldn't figure out why the asset browser wasn't working. It was because I had AWS_URL pointing to the Glide bucket, not the assets bucket. Hope this helps someone else.